azfunctions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
azure-functions==1.7.2
h5netcdf==0.11.0
netCDF4==1.5.7
pandas==1.3.3
psycopg2-binary==2.9.1 # NB. psycopyg2 requires libpq which is not available otherwise
shapely==1.7.1
xarray==0.19.0
